+++ sdmx/__init__.py
@@ -2,18 +2,19 @@ import logging
 
 import pkg_resources
 
-from sdmx.api import Request, read_url
+from sdmx.client import Client, Request, read_url
 from sdmx.reader import read_sdmx
 from sdmx.source import add_source, list_sources
 from sdmx.util import Resource
 from sdmx.writer import to_pandas, to_xml
 
 __all__ = [
+    "Client",
     "Request",
     "Resource",
     "add_source",
     "list_sources",
-    "logger",
+    "log",
     "read_sdmx",
     "read_url",
     "to_pandas",
@@ -29,15 +30,6 @@ except Exception:
 
 
 #: Top-level logger.
-logger = logging.getLogger(__name__)
-
-
-def _init_logger():
-    handler = logging.StreamHandler()
-    fmt = "{asctime} {name} - {levelname}: {message}"
-    handler.setFormatter(logging.Formatter(fmt, style="{"))
-    logger.addHandler(handler)
-    logger.setLevel(logging.ERROR)
-
-
-_init_logger()
+#:
+#: .. versionadded:: 0.4
+log = logging.getLogger(__name__)